Types

AstroNbodySim.CoarseMesh

AstroNbodySim.DS

AstroNbodySim.Delete

AstroNbodySim.DirectSum

AstroNbodySim.Euler

AstroNbodySim.FDM

AstroNbodySim.FEM

AstroNbodySim.FFT

AstroNbodySim.FVM

AstroNbodySim.GravityModel

AstroNbodySim.Leapfrog

AstroNbodySim.MHD

AstroNbodySim.ML

AstroNbodySim.MOND1983Milgrom

AstroNbodySim.Newton

AstroNbodySim.OutboundLimiter

AstroNbodySim.OutputInfo

AstroNbodySim.PhysicsInfo

AstroNbodySim.SPH

AstroNbodySim.SimConfig

AstroNbodySim.Simulation

AstroNbodySim.StreamInfo

AstroNbodySim.TimeInfo

AstroNbodySim.Tree

AstroNbodySim.TreeSimConfig

AstroNbodySim.VisualizationInfo

AstroNbodySim.Simulation — Type

struct Simulation{D}

config::Any

id::Pair{Int64, Int64}

pids::Vector{Int64}

simdata::Any

timeinfo::Any

::TimeInfo

outputinfo::Any

::OutputInfo

loginfo::Any

::LogInfo

physics::Any

::PhysicsInfo

stream::Any

::StreamInfo

visinfo::Any

::VisualizationInfo

buffer::Any

::Buffer

bgforce::Vector{Function}

bgpotential::Vector{Function}

AstroNbodySim.SimConfig — Type

struct SimConfig{F, U}

name::String

author::String

daytime::Dates.DateTime

floattype::Any

Numeric type of float numbers

units::Any

Prefered units in simulation. Default is uAstro. See PhysicalParticles: uAstro, uSI, uGadget2, uCGS. To run without units, use nothing

ZeroValues::Any

ZeroValue. Pre-constructed zero values for different types to simplify function arguments

constants::Any

Constant. Physical constants

loggingmode::Any

Choose how to display status of simulation. Supported: NormalMode, ProgressMode, SilentMode. Default is ProgressMode displaying progress bars

time::Any

TimeConfig

output::Any

OutputConfig

solver::Any

SolverConfig

grav::Any

GravityConfig

device::Any

DeviceConfig

Examples

SimConfig(; GravitySolver = Tree())
SimConfig(; device = GPU(), TimeStep = 1.0e-5u"Gyr")
SimConfig(; units = uGadget2)
SimConfig(; units = nothing, ForceSofteningTable = [0.01 for i in 1:6])
SimConfig(; TimeEnd = 1.0u"Gyr", OutputDir = "Test/Dir")

Missing docstring.

Missing docstring for LogInfo. Check Documenter's build log for details.

AstroNbodySim.StreamInfo — Type

mutable struct StreamInfo

loggingio::IOStream

timerio::IOStream

analyserio::IOStream

AstroNbodySim.TimeInfo — Type

mutable struct TimeInfo{T<:Number, I<:Integer}

dt::Number

Time interval between neighbor time steps

system_time_int::Integer

system_time_float::Number

last_system_time_int::Integer

last_system_time_float::Number

next_output_time_int::Integer

next_output_time_float::Number

redshift::Float64

scalefactor::Float64

min_endstep::Integer

stepcount::Integer

AstroNbodySim.OutputInfo — Type

mutable struct OutputInfo

snapshotcount::Int64

syncflag::Bool

AstroNbodySim.PhysicsInfo — Type

mutable struct PhysicsInfo

NumForceUpdateSinceLast::Int64

AstroNbodySim.VisualizationInfo — Type

mutable struct VisualizationInfo

progress::ProgressMeter.Progress

PlotData::Any

resolution::Any

fig::Any

Realtime::Bool

RenderTime::Float64

last_plot_time::Float64

xlims::Any

ylims::Any

zlims::Any

markersize::Float64

AstroNbodySim.TreeSimConfig — Type

struct TreeSimConfig

Controls accuracy of tree method

TreeOpenAngle::Float64

If the view angle of tree node that relative to the sink point is too large, open this node. In radian unit. Default is 0.1

ErrTolAcc::Float64

While opening the tree node, take the last acceleration into account. Improves accuracy at high redshift cosmology. Default is 0.025

Missing docstring.

Missing docstring for OctreeData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Buffer. Check Documenter's build log for details.

Trait types

AstroNbodySim.DirectSum — Type

Direct Summation Method

AstroNbodySim.Tree — Type

Peano-Hilbert Space Filling Octree Method

AstroNbodySim.FDM — Type

Finite Differencing Method

AstroNbodySim.FFT — Type

Fast Fourier Transform

AstroNbodySim.ML — Type

Machine Learning

AstroNbodySim.SPH — Type

Smoothed Particle Hydrodynamics

AstroNbodySim.MHD — Type

Magnetohydrodynamics

AstroNbodySim.FEM — Type

Finite Element Method

AstroNbodySim.FVM — Type

Finite Volume Method

AstroNbodySim.GravityModel — Type

Gravity model. Supported: Newton, MOND1983Milgrom, QUMOND

AstroNbodySim.Newton — Type

Traditional Newtonian gravity

AstroNbodySim.MOND1983Milgrom — Type

Milgrom 1983 formula of MOND

Missing docstring.

Missing docstring for QUMOND. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LoggingMode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NormalMode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ProgressMode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SilentMode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TimeIntegration. Check Documenter's build log for details.

AstroNbodySim.Euler — Type

1st-order explicit Euler time integration

AstroNbodySim.Leapfrog — Type

Leapfrog time integration

Missing docstring.

Missing docstring for GPUAlgorithm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AllPairs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Tiled. Check Documenter's build log for details.

AstroNbodySim.OutboundLimiter — Type

`::OutboundLimiter`. Choose how to handle particles out of the non-periodic simulation box.
Supported:
    - `Delete`: delete outbound particles
    - `DS`: use direct summation method to compute forces 
    - `CoarseMesh`: Construct a coarse mesh to overlap all particles

AstroNbodySim.Delete — Type

Delete outbound particles if they run out of the non-periodic simulation box

AstroNbodySim.DS — Type

Compute forces using direct summation method if the particles run out of the non-periodic simulation box

AstroNbodySim.CoarseMesh — Type

Construct a coarse mesh to overlap all particles for non-periodic boundary conditions